Transfer fetch intervals to respective tables (LR and JEV)
